# openapi.yaml
openapi: 3.0.3
info:
  title: Sistem Pendataan Inventaris Laboratorium REST API
  description: |
    API Spesifikasi untuk Sistem Pendataan Inventaris Laboratorium (Sistem Pendataan Inventaris Laboratorium) 
    menggunakan Spring Boot, JWT Authentication, dan Role-Based Access Control (RBAC).
    
    ### Peran Pengguna (Roles):
    - **ADMIN**: Memiliki akses penuh ke seluruh sistem termasuk CRUD kategori, barang, laporan, dan aksi peminjaman.
    - **PETUGAS**: Memiliki hak akses serupa dengan ADMIN untuk operasional sehari-hari (CRUD kategori/barang, persetujuan peminjaman, laporan).
    - **MAHASISWA**: Hanya dapat melihat kategori/barang, melakukan pengajuan peminjaman barang untuk diri sendiri, dan melihat riwayat peminjaman miliknya sendiri.
    
    ### Alur Autentikasi:
    1. Lakukan registrasi melalui `/api/auth/register` (opsional jika menggunakan akun seeder bawaan).
    2. Lakukan login melalui `/api/auth/login` untuk mendapatkan token JWT.
    3. Masukkan token tersebut dalam Header HTTP: `Authorization: Bearer <JWT_TOKEN>` untuk mengakses endpoint yang dilindungi.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Development Server
paths:
...
  /api/peminjaman:
    get:
      tags:
        - Transaksi Peminjaman
      summary: Dapatkan Riwayat/Daftar Transaksi Peminjaman
      description: |
        Mengambil daftar pengajuan peminjaman barang dengan paginasi.
        - **ADMIN** dan **PETUGAS** dapat melihat seluruh transaksi dari semua pengguna.
        - **MAHASISWA** hanya dapat melihat riwayat peminjaman miliknya sendiri.
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: Nomor halaman (0-indexed)
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Jumlah data per halaman
          required: false
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          description: Kolom untuk pengurutan
          required: false
          schema:
            type: string
            default: idPeminjaman
        - name: sortDir
          in: query
          description: Arah pengurutan (ASC atau DESC)
          required: false
          schema:
            type: string
            enum: [ASC, DESC]
            default: DESC
      responses:
        '200':
          description: Berhasil memuat daftar transaksi peminjaman
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseListPeminjamanResponse'
        '401':
          description: Token tidak valid

    post:
      tags:
        - Transaksi Peminjaman
      summary: Ajukan Peminjaman Barang
      description: |
        Mengajukan peminjaman satu atau beberapa barang inventaris sekaligus.
        - Status peminjaman default diset ke `MENUNGGU` (Pending).
        - Sistem belum memotong stok saat pengajuan dibuat.
        - Tanggal pinjam harus hari ini atau di masa mendatang.
        - Dapat diakses oleh ADMIN, PETUGAS, dan MAHASISWA.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PeminjamanRequest'
      responses:
        '201':
          description: Pengajuan peminjaman berhasil dibuat, menunggu persetujuan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponsePeminjamanResponse'
        '400':
          description: Validasi gagal / Tanggal pinjam melanggar aturan / ID barang salah
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseValidationError'
  /api/peminjaman/{id}/approve:
    put:
      tags:
        - Transaksi Peminjaman
      summary: Setujui Peminjaman Barang
      description: |
        Menyetujui transaksi peminjaman yang berstatus `MENUNGGU`.
        - Stok tersedia (`jumlahTersedia`) pada barang-barang terkait akan dikurangi secara otomatis.
        - Memastikan jumlah stok mencukupi sebelum mengurangi stok agar tidak negatif.
        - Hanya dapat diakses oleh **ADMIN** dan **PETUGAS**.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: ID Transaksi Peminjaman
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Transaksi peminjaman berhasil disetujui, stok berkurang
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponsePeminjamanResponse'
        '400':
          description: Peminjaman tidak berstatus MENUNGGU / Stok barang tidak mencukupi untuk disetujui
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseString'
        '404':
          description: Transaksi peminjaman tidak ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseString'
        '403':
          description: Hak akses ditolak
  /api/peminjaman/{id}/reject:
    put:
      tags:
        - Transaksi Peminjaman
      summary: Tolak Peminjaman Barang
      description: |
        Menolak pengajuan transaksi peminjaman yang berstatus `MENUNGGU`.
        - Mengubah status transaksi menjadi `DITOLAK`.
        - Tidak mempengaruhi stok barang.
        - Hanya dapat diakses oleh **ADMIN** dan **PETUGAS**.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: ID Transaksi Peminjaman
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Pengajuan peminjaman ditolak
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponsePeminjamanResponse'
        '400':
          description: Peminjaman tidak berstatus MENUNGGU (misal sudah disetujui atau dibatalkan sebelumnya)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseString'
        '404':
          description: Transaksi tidak ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseString'
        '403':
          description: Hak akses ditolak
  /api/peminjaman/{id}/kembalikan:
    put:
      tags:
        - Transaksi Peminjaman
      summary: Kembalikan Barang Peminjaman
      description: |
        Memproses pengembalian barang atas transaksi peminjaman yang berstatus `DISETUJUI`.
        - Stok barang tersedia (`jumlahTersedia`) akan otomatis dipulihkan / ditambah kembali.
        - Kondisi akhir masing-masing barang saat dikembalikan (BAIK, RUSAK_RINGAN, RUSAK_BERAT) dapat dicatat melalui request body (opsional).
        - Mengubah status transaksi menjadi `DIKEMBALIKAN`.
        - Hanya dapat diakses oleh **ADMIN** dan **PETUGAS**.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: ID Transaksi Peminjaman yang akan dikembalikan barangnya
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: false
        description: Detail kondisi pengembalian barang. Jika kosong, kondisi dianggap BAIK.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PengembalianRequest'
      responses:
        '200':
          description: Barang peminjaman berhasil dikembalikan, stok dipulihkan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponsePeminjamanResponse'
        '400':
          description: Peminjaman tidak berstatus DISETUJUI / Sudah dikembalikan sebelumnya
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseString'
        '404':
          description: Transaksi tidak ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseString'
        '403':
          description: Hak akses ditolak
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Masukkan token JWT Anda dalam format `Bearer <token>` pada header Otorisasi HTTP.
  schemas:
...
    PeminjamanRequest:
      type: object
      required:
        - tglPinjam
        - tglKembali
        - details
      properties:
        tglPinjam:
          type: string
          format: date
          example: "2026-05-18"
          description: Tanggal peminjaman barang (harus hari ini atau masa mendatang)
        tglKembali:
          type: string
          format: date
          example: "2026-05-25"
          description: Tanggal rencana pengembalian barang
        catatan:
          type: string
          example: Untuk keperluan praktikum mata kuliah Mikroprosesor minggu ini.
        details:
          type: array
          items:
            $ref: '#/components/schemas/DetailPeminjamanRequest'
    DetailPeminjamanRequest:
      type: object
      required:
        - idBarang
        - jumlah
      properties:
        idBarang:
          type: integer
          format: int64
          example: 1
        jumlah:
          type: integer
          minimum: 1
          example: 2
    PengembalianRequest:
      type: object
      required:
        - details
      properties:
        details:
          type: array
          items:
            $ref: '#/components/schemas/PengembalianDetailRequest'
    PengembalianDetailRequest:
      type: object
      required:
        - idDetail
        - kondisiKembali
      properties:
        idDetail:
          type: integer
          format: int64
          example: 1
          description: ID Detail peminjaman barang yang dikembalikan
        kondisiKembali:
          $ref: '#/components/schemas/KondisiBarang'

    # Enums
    Role:
      type: string
      enum:
        - ADMIN
        - PETUGAS
        - MAHASISWA
      example: ADMIN
    KondisiBarang:
      type: string
      enum:
        - BAIK
        - RUSAK_RINGAN
        - RUSAK_BERAT
      example: BAIK
    StatusPeminjaman:
      type: string
      enum:
        - MENUNGGU
        - DISETUJUI
        - DITOLAK
        - DIKEMBALIKAN
      example: MENUNGGU

    # Wrapper Response Objects
    WebResponseString:
      type: object
      properties:
        code:
          type: integer
          example: 200
        status:
          type: string
          example: OK
        message:
          type: string
          example: Operasi berhasil!
        data:
          type: string
          nullable: true
          example: null
    WebResponseValidationError:
      type: object
      properties:
        code:
          type: integer
          example: 400
        status:
          type: string
          example: BAD REQUEST
        message:
          type: string
          example: Validation failed
        data:
          type: object
          additionalProperties:
            type: string
          example:
            email: Format email tidak valid!
            password: Password minimal terdiri dari 6 karakter!
    PagingResponse:
      type: object
      properties:
        currentPage:
          type: integer
          example: 0
        totalPages:
          type: integer
          example: 1
        size:
          type: integer
          example: 10
        totalElements:
          type: integer
          format: int64
          example: 5

    # Entity Response Objects
    UserResponse:
      type: object
      properties:
        idUser:
          type: integer
          format: int64
          example: 1
        nama:
          type: string
          example: Ahmad Mahasiswa
        email:
          type: string
          format: email
          example: mahasiswa@lab.com
        role:
          $ref: '#/components/schemas/Role'
        createdAt:
          type: string
          format: date-time
          example: "2026-05-18T08:50:00Z"
...
    DetailPeminjamanResponse:
      type: object
      properties:
        idDetail:
          type: integer
          format: int64
          example: 1
        idBarang:
          type: integer
          format: int64
          example: 1
        namaBarang:
          type: string
          example: Digital Multimeter Sanwa CD800a
        kodeBarang:
          type: string
          example: BRG-EL-001
        jumlah:
          type: integer
          example: 2
        kondisiKembali:
          anyOf:
            - $ref: '#/components/schemas/KondisiBarang'
            - type: string
              nullable: true
          description: Kondisi barang saat dikembalikan. Berstatus `null` selama status pinjam belum DIKEMBALIKAN.
    PeminjamanResponse:
      type: object
      properties:
        idPeminjaman:
          type: integer
          format: int64
          example: 1
        peminjam:
          $ref: '#/components/schemas/UserResponse'
        petugas:
          anyOf:
            - $ref: '#/components/schemas/UserResponse'
            - type: object
              nullable: true
          description: Petugas/Admin yang melakukan verifikasi transaksi. Bernilai `null` jika status peminjaman masih MENUNGGU.
        tglPinjam:
          type: string
          format: date
          example: "2026-05-18"
        tglKembali:
          type: string
          format: date
          example: "2026-05-25"
        status:
          $ref: '#/components/schemas/StatusPeminjaman'
        catatan:
          type: string
          example: Untuk keperluan praktikum mata kuliah Mikroprosesor minggu ini.
        details:
          type: array
          items:
            $ref: '#/components/schemas/DetailPeminjamanResponse'
    WebResponsePeminjamanResponse:
      type: object
      properties:
        code:
          type: integer
          example: 201
        status:
          type: string
          example: CREATED
        message:
          type: string
          example: Pengajuan peminjaman berhasil dibuat, menunggu persetujuan petugas!
        data:
          $ref: '#/components/schemas/PeminjamanResponse'
    WebResponseListPeminjamanResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        status:
          type: string
          example: OK
        message:
          type: string
          example: Daftar transaksi peminjaman berhasil dimuat!
        data:
          type: array
          items:
            $ref: '#/components/schemas/PeminjamanResponse'
        paging:
          $ref: '#/components/schemas/PagingResponse'
...